+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.1: 2025-07-03
 
 * Add loved feature to Popular section in artist views
+* Implement loved tracks feature to album views
 
 ### 1.1.0: 2025-07-03